Remove unused npm package dependencies
The old dependencies in package.json cause security warnings in npm.
The package.json file was originally copied from the Cupper project
https://github.com/ThePacielloGroup/cupper.
The Cupper project has not been maintained for a few years.
However, this A+ accessibility audit is actually directly built
with Hugo using a Docker container. Therefore, the dependencies
listed for Cupper are irrelevant. Removing the dependencies from
package.json silences the security warnings.